+++ examples/demo/config.tsx
@@ -202,11 +202,13 @@ export default (skin: string) => {
       title: "Are you sure delete this rule?",
       okText: "Yes",
       okType: "danger",
+      cancelText: "Cancel"
     },
     removeGroupConfirmOptions: {
       title: "Are you sure delete this group?",
       okText: "Yes",
       okType: "danger",
+      cancelText: "Cancel"
     },
   };
 
@@ -238,6 +240,7 @@ export default (skin: string) => {
     // showLabels: true,
     maxNesting: 5,
     canLeaveEmptyGroup: true,
+    shouldCreateEmptyGroup: false,
     showErrorMessage: true,
     // renderField: (props) => <FieldCascader {...props} />,
     // renderOperator: (props) => <FieldDropdown {...props} />,
@@ -349,7 +352,7 @@ export default (skin: string) => {
           type: "number",
           fieldSettings: {
             min: 1990,
-            max: 2020,
+            max: 2021,
           },
           valueSources: ["value"],
         }