+++ modules/config/material/index.js
@@ -1,7 +1,8 @@
 import MaterialWidgets from "../../components/widgets/material";
-import BasicConfig from "../basic";
+import BasicConfig, {stringifyForDisplay} from "../basic";
 import React from "react";
 import {SqlString} from "../../utils/sql";
+
 
 const {
   MaterialBooleanWidget,
@@ -108,7 +109,7 @@ const widgets = {
       { label: "Number to", placeholder: "Enter number to" },
     ],
     formatValue: (val, fieldDef, wgtDef, isForDisplay) => {
-      return isForDisplay ? val : JSON.stringify(val);
+      return isForDisplay ? stringifyForDisplay(val) : JSON.stringify(val);
     },
     sqlFormatValue: (val, fieldDef, wgtDef, op, opDef) => {
       return SqlString.escape(val);